ss_twr_resp now properly sends transmission timestamp & ss_twr_init now properly receives it

+++ dwm1001-examples-master/examples/ss_twr_init_int/ss_init_main.h
@@ -13,7 +13,7 @@
  */
 
 /* Declaration of static functions. */
-static void resp_msg_get_ts(uint8 *ts_field, uint32 *ts);
+static void resp_msg_get_ts(uint8 *ts_field, uint64 *ts);
 
 int ss_init_run(void);
 
@@ -25,6 +25,6 @@ void rx_err_cb(const dwt_cb_data_t *cb_data);
 
 void tx_conf_cb(const dwt_cb_data_t *cb_data);
 
-static void resp_msg_get_ts(uint8 *ts_field, uint32 *ts);
+static void resp_msg_get_ts(uint8 *ts_field, uint64 *ts);
 
 void ss_initiator_task_function (void * pvParameter);
